feat: déplace la logique de traitement des contributions vers le contrôleur
